Move useSelect to shared util directory
